Agentic_Physics_Research_Lab/server/requirements.txt
fastapi
uvicorn
openai
anthropic
python-dotenv
pytest
httpx
wolframalpha
wolframclient
